+++ src/main/java/chess/dto/piece/PieceDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.piece;
 
 import chess.domain.piece.Piece;
 import chess.domain.team.Team;
